Seed Sample Meeting Data For Development
Something is screwy with our environmental variables in insertTestData.ts. For the moment we're using dotEnv and a separate .env file in the data directory. Before this is all over, we'll want to figure this out, because we don't want to keep a separate .env file just for this function.