Pass Meeting Data and Titles to Modals
In order to update our meetings with any possible yearly meetings, accessibilities, etc., we need all that data available in the modal.

And, in order to have our selections populate dynamically, we also need to provide all the names for each state - and soon for yearly meetings, accessibilities, etc.

We're passing data down deeply enough at this point that it does raise the question of whether or not a state management tool beyond React state might be useful.