Handle Update Meeting (part 1)
We're attempting to make a PUT request to update a meeting's attributes. Unfortunately, we're currently just seeing a preflight OPTIONS request being made, without being followed up with a PUT request.

Once we resolve this, we'll also need to handle properly updating state in ModifyMeetingModal, so that we have the proper data to update our meeting record.