Update README With Postgres Setup Info
Setting up Postgres was not well-documented, and there are some issues with db-migrate that needed explanation.

If we have development going on in a Linux or Windows environment, we'll probably need to add documentation for that, too. For the time being, we provide detailed instructions for setting up psql on Mac.