Use Pool Rather Than Single Client for Database Requests
Before, we were creating a single client, and that worked for a small number of test records. Now that we're dealing with thousands of records at once, however, we need to be able to support numerous concurrent clients.

Node-Postgres lets us create a pool and then draw clients from the pool. Whenever we're done with a spawned client, we can .release it and free it up to handle more requests.