Insert Development Data into SQL Database
* Create insertDevelopmentData.js which uses the node-postgres pool rather than a single client (this allows us to handle a lot more records)
* Add language and school_time to the Meeting schema

Outstanding Problems:

We've uncovered some serious issues here with data integrity. Some of the Meeting records were scraped improperly due to typos and the like in the source data. We need to handle for this, and probably just do some manual corrections at this point, in some cases.

Additionally to these typos, we seem to be doing something wrong in our transformation pipeline. Here are the following known issues at this time:

* Yearly Meetings consist of approximately half a dozen Monthly Meetings
* Many meetings don't have a state and so are assigned to "0" - additionally, most meetings just aren't assigned properly to a state
* "American Sign Language" as an accessibility option seems to be broken (doesn't update when selected)